LayoutTests/fast/dom/webtiming-expected.txt

This test checks that all of the Web Timing attributes are available and have reasonable values in the right order.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL timing.navigationStart should be >= oneHourAgoUTC. Was undefined (of type undefined).
FAIL timing.unloadEventEnd should be >= timing.navigationStart. Was undefined (of type undefined).
FAIL timing.loadEventStart should be 0 (of type number). Was undefined (of type undefined).
FAIL timing.loadEventEnd should be 0 (of type number). Was undefined (of type undefined).
FAIL timing.navigationStart should be >= oneHourAgoUTC. Was undefined (of type undefined).
FAIL timing.unloadEventEnd should be >= timing.navigationStart. Was undefined (of type undefined).
FAIL timing.responseEnd should be >= timing.unloadEventEnd. Was undefined (of type undefined).
FAIL timing.loadEventStart should be >= timing.responseEnd. Was undefined (of type undefined).
FAIL timing.loadEventEnd should be 0 (of type number). Was undefined (of type undefined).
FAIL timing.navigationStart should be >= oneHourAgoUTC. Was undefined (of type undefined).
FAIL timing.unloadEventEnd should be >= timing.navigationStart. Was undefined (of type undefined).
FAIL timing.responseEnd should be >= timing.unloadEventEnd. Was undefined (of type undefined).
FAIL timing.loadEventStart should be >= timing.responseEnd. Was undefined (of type undefined).
FAIL timing.loadEventEnd should be >= timing.loadEventStart + 50. Was undefined (of type undefined).
PASS successfullyParsed is true

TEST COMPLETE
